+++ include/renderer.h
@@ -77,7 +77,7 @@ void drawKeyboard(bool tv);
 void textToFrame(int row, int column, const char* str);
 void lineToFrame(int column, uint32_t color);
 void boxToFrame(int lineStart, int lineEnd);
-void barToFrame(int line, int column, uint32_t width, float progress);
+void barToFrame(int line, int column, uint32_t width, double progress);
 void arrowToFrame(int line, int column);
 void checkmarkToFrame(int line, int column);
 void tabToFrame(int line, int column, char *label, bool active);
